# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "This is the OpenAPI specification for your project."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/account": {
      "post": {
        "operationId": "post-account",
        "summary": "Create account\r",
        "description": "Creates a new customer account. Returns the newly created account",
        "tags": [
          "Account"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ssn": {
                    "type": "string",
                    "nullable": false
                  },
                  "name": {
                    "type": "string",
                    "nullable": false
                  },
                  "email": {
                    "type": "string",
                    "nullable": false
                  },
                  "address": {
                    "type": "string",
                    "nullable": false
                  },
                  "phone_nr": {
                    "type": "string",
                    "nullable": false
                  },
                  "password": {
                    "type": "string",
                    "nullable": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "get": {
        "operationId": "get-account",
        "summary": "Get account\r",
        "description": "TODO: Add session management",
        "tags": [
          "Account"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "put": {
        "operationId": "put-account",
        "summary": "Update account\r",
        "description": "TODO: Add session management",
        "tags": [
          "Account"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete-account",
        "summary": "Delete account\r",
        "description": "TODO: Add session management",
        "tags": [
          "Account"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/admin/product": {
      "post": {
        "operationId": "post-admin-product",
        "summary": "Create a new product",
        "description": "",
        "tags": [
          "Admin"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "put": {
        "operationId": "put-admin-product",
        "summary": "Update product information",
        "description": "",
        "tags": [
          "Admin"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/admin/product/[id]": {
      "delete": {
        "operationId": "delete-admin-product-[id]",
        "summary": "Delete product",
        "description": "",
        "tags": [
          "Admin"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/product": {
      "get": {
        "operationId": "get-product",
        "summary": "Get all products",
        "description": "",
        "tags": [
          "Product"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/product/[id]": {
      "get": {
        "operationId": "get-product-[id]",
        "summary": "Get a product by id",
        "description": "",
        "tags": [
          "Product"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/test": {
      "get": {
        "operationId": "get-test",
        "summary": "",
        "description": "",
        "tags": [
          "Test"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  }
}